Feat: Troca o dataTypes de price para que retorne number ao invés de string

+++ back-end/src/database/migrations/20230504160228-create-products.js
@@ -15,7 +15,7 @@ module.exports = {
 
       price: {
         allowNull: true,
-        type: Sequelize.DECIMAL
+        type: Sequelize.DOUBLE(4, 2),
       },
       
       urlImage: {
